+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -9,6 +9,12 @@ CHANGELOG
  * Added `ControllerTrait::isFormValid()`
  * Added an `help_html` form option to display the `help` text as HTML
 
+ * [BC Break] When using Messenger, the default transport changed from
+   using Symfony's serializer service to use `PhpSerializer`, which uses
+   PHP's native `serialize()` and `unserialize()` functions. To use the
+   original serialization method, set the `framework.messenger.serializer.id`
+   config option to `messenger.transport.symfony_serializer`.
+
 4.2.0
 -----